feat(comp): refactor most views
Most views have been refactored now, the ones directly accessible via the row of menu buttons at least. Finally, I will set the login page to be the first page visited, and hide the navbar if a user is not logged in.
