+++ src/main/java/org/trainer/interval_trainer/controller/SettingsController.java
@@ -13,105 +13,7 @@ import java.io.IOException;
 
 public class SettingsController {
     @FXML
-    private Label welcomeText;
-    @FXML
-    private Label header;
-    @FXML
-    private Button homeButton;
-    @FXML
-    private Button createRoutinesButton;
-    @FXML
-    private Button searchButton;
-    @FXML
-    private Button profileButton;
-    @FXML
-    private Button settingsButton;
-
-    @FXML
     protected void initialize() {
-        welcomeText.setText("Settings");
-        header.setText("Interval Training App");
+
     }
-
-    @FXML
-    protected void onHomeButtonClick() throws IOException {
-        // Get the current scene
-        Scene currentScene = homeButton.getScene();
-
-        // Check if the current scene is already the "create-routine-view.fxml"
-        if (currentScene.getRoot().getId().equals("helloRoot")) {
-            // Already on the new routine page, no need to reload
-            return;
-        }
-
-        Stage stage = (Stage) homeButton.getScene().getWindow();
-        FXMLLoader fxmlLoader = new FXMLLoader(HelloApplication.class.getResource("hello-view.fxml"));
-        Scene scene = new Scene(fxmlLoader.load(), HelloApplication.WIDTH, HelloApplication.HEIGHT);
-        stage.setScene(scene);
-    }
-
-
-    @FXML
-    protected void onCreateRoutinesButtonClick() throws IOException {
-        // Get the current scene
-        Scene currentScene = createRoutinesButton.getScene();
-
-        // Check if the current scene is already the "create-routine-view.fxml"
-        if (currentScene.getRoot().getId().equals("createRoutineRoot")) {
-            // Already on the new routine page, no need to reload
-            return;
-        }
-        Stage stage = (Stage) createRoutinesButton.getScene().getWindow();
-        FXMLLoader fxmlLoader = new FXMLLoader(HelloApplication.class.getResource("create-routine-view.fxml"));
-        Scene scene = new Scene(fxmlLoader.load(), HelloApplication.WIDTH, HelloApplication.HEIGHT);
-        stage.setScene(scene);
-    }
-
-    @FXML
-    protected void onSearchButtonClick() throws IOException {
-        // Get the current scene
-        Scene currentScene = searchButton.getScene();
-
-        // Check if the current scene is already the "create-routine-view.fxml"
-        if (currentScene.getRoot().getId().equals("searchRoot")) {
-            // Already on the new routine page, no need to reload
-            return;
-        }
-        Stage stage = (Stage) searchButton.getScene().getWindow();
-        FXMLLoader fxmlLoader = new FXMLLoader(HelloApplication.class.getResource("search-view.fxml"));
-        Scene scene = new Scene(fxmlLoader.load(), HelloApplication.WIDTH, HelloApplication.HEIGHT);
-        stage.setScene(scene);
-    }
-    @FXML
-    protected void onProfileButtonClick() throws IOException {
-        // Get the current scene
-        Scene currentScene = profileButton.getScene();
-
-        // Check if the current scene is already the "create-routine-view.fxml"
-        if (currentScene.getRoot().getId().equals("profileRoot")) {
-            // Already on the new routine page, no need to reload
-            return;
-        }
-        Stage stage = (Stage) profileButton.getScene().getWindow();
-        FXMLLoader fxmlLoader = new FXMLLoader(HelloApplication.class.getResource("profile-view.fxml"));
-        Scene scene = new Scene(fxmlLoader.load(), HelloApplication.WIDTH, HelloApplication.HEIGHT);
-        stage.setScene(scene);
-    }
-
-    @FXML
-    protected void onSettingsButtonClick() throws IOException {
-        // Get the current scene
-        Scene currentScene = settingsButton.getScene();
-
-        // Check if the current scene is already the "create-routine-view.fxml"
-        if (currentScene.getRoot().getId().equals("settingsRoot")) {
-            // Already on the new routine page, no need to reload
-            return;
-        }
-        Stage stage = (Stage) profileButton.getScene().getWindow();
-        FXMLLoader fxmlLoader = new FXMLLoader(HelloApplication.class.getResource("settings-view.fxml"));
-        Scene scene = new Scene(fxmlLoader.load(), HelloApplication.WIDTH, HelloApplication.HEIGHT);
-        stage.setScene(scene);
-    }
-
 }
