+++ src/main/resources/org/trainer/interval_trainer/profile-view.fxml
@@ -12,24 +12,6 @@
 <?import javafx.scene.image.ImageView?>
 <?import javafx.scene.image.Image?>
 <BorderPane fx:id="profileRoot" xmlns:fx="http://javafx.com/fxml" fx:controller="org.trainer.interval_trainer.controller.ProfileController" style="-fx-font-family: 'Arial'">
-    <!--    Header-->
-    <top>
-        <HBox alignment="TOP_LEFT" spacing="10.0" style="-fx-border-color: black; -fx-border-width: 1px;">
-
-            <padding>
-                <Insets bottom="10.0" left="10.0" right="10.0" top="10.0"/>
-            </padding>
-            <!-- Default Logo Image -->
-            <ImageView fitWidth="50" fitHeight="50">
-                <image>
-                    <Image url="@logo.png" />
-                </image>
-            </ImageView>
-            <!-- App Name Label -->
-            <Label fx:id="header" style="-fx-font-size: 35; -fx-font-weight: bold" />
-        </HBox>
-
-    </top>
     <center>
         <VBox alignment="TOP_CENTER" spacing="20.0">
             <padding>
@@ -46,19 +28,5 @@
             <Label fx:id="fitnessGoalLabel" style="-fx-font-size: 18;" wrapText="true" />
 
         </VBox>
-
-
     </center>
-    <bottom>
-        <HBox alignment="CENTER" spacing="20.0" style="-fx-border-color: black; -fx-border-width: 1px;">
-            <padding>
-                <Insets bottom="20.0" left="20.0" right="20.0" top="20.0"/>
-            </padding>
-            <Button fx:id="homeButton" text="Home" onAction="#onHomeButtonClick"/>
-            <Button fx:id="createRoutinesButton" text="Create Routines" onAction="#onCreateRoutinesButtonClick"/>
-            <Button fx:id="searchButton" text="Search" onAction="#onSearchButtonClick"/>
-            <Button fx:id="profileButton" text="Profile" onAction="#onProfileButtonClick"/>
-            <Button fx:id="settingsButton" text="Settings" onAction="#onSettingsButtonClick"/>
-        </HBox>
-    </bottom>
 </BorderPane>
